% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_beta_values_all.R
\docType{data}
\name{betaVals_mat}
\alias{betaVals_mat}
\title{Annotated CPG Data Set}
\format{A matrix containing 356603 CPGs measured on 14 subjects. The rownames
   are the CPG IDs, and the column names indicate the sample IDs of the
   subjects.}
\source{
Calculated via the \code{1_Aclust_data_import.R} script in the
   \code{old_scripts} sub-directory of the \code{inst} directory.
}
\usage{
betaVals_mat
}
\description{
A beta value matrix for selected methylation samples from a 450k
   methylation array. These beta values are trimmed to range from 0.05 to
   0.95.
}
\keyword{datasets}
