% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Build_DMR-Overlap_List.R
\name{BuildOverlaps}
\alias{BuildOverlaps}
\title{Build a List of DMR-Overlap Lists}
\usage{
BuildOverlaps(bestResultsDir, delta = c(0.025, 0.05, 0.1, 0.15, 0.2, 0.3,
  0.4), seed = c(100, 210, 330, 450, 680), CPGs_df = cpgLocation_df,
  min.cpgs = 5)
}
\arguments{
\item{bestResultsDir}{The name of the directory where the method results from
the best-performing parameter settings are stored. For the full design we
have included (\code{delta = c(0.025, 0.05, 0.1, 0.15, 0.2, 0.3, 0.4)} and
\code{seed = c(100, 210, 330, 450, 680)}), this directory should contain
35 \code{.RDS} files per method.}

\item{delta}{A treatment size corresponding to one of the simulations with
completed results files in the \code{bestResultsDir} directory.}

\item{seed}{A seed value corresponding to one of the simulations with
completed results files in the \code{bestResultsDir} directory.}

\item{CPGs_df}{An annotation table that indicates locations of CpGs.
This data frame has CPG IDs as the rows with matching chromosome and
location info in the columns. Specifically, the columns are: \code{ILMNID}
 - the CPG ID; \code{chr} - the chromosome label; and \code{MAPINFO} -
the chromosome location. An example is given in the \code{cpgLocation_df}
data set. This data set is only necessary if the results directory
contains Comb-p results with the specified \code{delta} and \code{seed}
values.}

\item{min.cpgs}{The minimum number of CPGs before we consider a result
significant. Defaults to 5. This argument is only required if the results
directory contains Comb-p results with the specified \code{delta} and
\code{seed} values.}
}
\value{
A list of DMR overlaps to pass to the
   \code{\link[ChIPpeakAnno]{makeVennDiagram}} function.
}
\description{
Given a directory of best-performing results files from one of
   the simulation functions (\code{\link{WriteDMRcateResults}},
   \code{\link{WriteProbeLassoResults}},
   \code{\link{WriteBumphunterResults}}, or results from the \code{Comb-p}
   method in \code{Python}), import the raw data files and DMR overlap lists
   via the \code{\link[GenomicRanges]{GRanges}} and
   \code{\link[IRanges]{IRanges}} functions.
}
\details{
This function is called internally by the \code{\link{PlotOverlaps}}
   function.
}
\examples{
  # Called internally by the PlotOverlaps() function.
}
\keyword{internal}
