% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Build_Precision-Recall_Curve_List.R
\name{BuildPRcurve}
\alias{BuildPRcurve}
\title{Build a List of Precision-Recall Curve Objects}
\usage{
BuildPRcurve(bestResultsDir, delta = c(0.025, 0.05, 0.1, 0.15, 0.2, 0.3, 0.4),
  seed = c(100, 210, 330, 450, 680), beta_mat = betaVals_mat,
  AclustCPG_df = startEndCPG_df, CPGs_df = cpgLocation_df, min.cpgs = 5)
}
\arguments{
\item{bestResultsDir}{The name of the directory where the method results from
the best-performing parameter settings are stored. For the full design we
have included (\code{delta = c(0.025, 0.05, 0.1, 0.15, 0.2, 0.3, 0.4)} and
\code{seed = c(100, 210, 330, 450, 680)}), this directory should contain
35 \code{.RDS} files per method.}

\item{delta}{A treatment size corresponding to one of the simulations with
completed results files in the \code{bestResultsDir} directory.}

\item{seed}{A seed value corresponding to one of the simulations with
completed results files in the \code{bestResultsDir} directory.}

\item{beta_mat}{A beta value matrix for selected methylation samples from a
450k methylation array with CPG IDs as the row names and sample IDs as
the column names. An example is given in the \code{betaVals_mat} data set.}

\item{AclustCPG_df}{A data frame of beta values and CpG information for
clusters of CpGs over a 450k methylation array. The rows correspond to the
CPGs. The columns have information on the cluster number, chromosome,
cluster start and end locations, and the beta values for each subject
grouped by some clinical indicator (e.g. case v. control). An example is
given in the \code{startEndCPG_df} data set.}

\item{CPGs_df}{An annotation table that indicates locations of CpGs.
This data frame has CPG IDs as the rows with matching chromosome and
location info in the columns. Specifically, the columns are: \code{ILMNID}
 - the CPG ID; \code{chr} - the chromosome label; and \code{MAPINFO} -
the chromosome location. An example is given in the \code{cpgLocation_df}
data set. This data set is only necessary if the results directory
contains Comb-p results with the specified \code{delta} and \code{seed}
values.}

\item{min.cpgs}{The minimum number of CPGs necessary to consider a result
significant. Defaults to 5. This argument is only required if the results
directory contains Comb-p results with the specified \code{delta} and
\code{seed} values.}
}
\value{
A list of PR-curve objects, to be plotted via the
   \code{\link{PlotPRCurve}} function.
}
\description{
Given a directory of best-performing results files from one of
   the simulation functions (\code{\link{WriteDMRcateResults}},
   \code{\link{WriteProbeLassoResults}},
   \code{\link{WriteBumphunterResults}}, or results from the \code{Comb-p}
   method in \code{Python}), import the raw data files and construct PR-curve
   objects via the \code{\link[PRROC]{pr.curve}} function.
}
\examples{
\dontrun{
   BuildPRcurve(
     bestResultsDir = "best_cases_results/",
     delta = 0.4,
     seed = 100
   )
}
}
