% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_clean_DMR_results.R
\name{CleanResults}
\alias{CleanResults}
\title{Clean and Join DMR Method Predictions to True Values}
\usage{
CleanResults(dmrResults_ls, Aclusters_df)
}
\arguments{
\item{dmrResults_ls}{A list of results, written by one of the functions
\code{\link{WriteDMRcateResults}}, \code{\link{WriteBumphunterResults}},
or \code{\link{WriteProbeLassoResults}}.}

\item{Aclusters_df}{A data frame of true cluster values, returned in the
\code{simAclusters} entry of the output of the \code{\link{SimulateData}}
function.}
}
\value{
A data frame with the predicted clusters joined to their true values.
   This data frame is to be immediately passed to the
   \code{\link{SummarizeResults}} function for analysis.
}
\description{
Given a results file, as written by one of the
   \code{Write*Results} functions, and the data frame of true cluster values,
   as returned by the \code{SimulateData}, join the predicted clusters to the
   true clusters.
}
\examples{
  # Called internally by the ProcessDMRcateResults(), ProcessCombpResults(),
  #   ProcessProbeLassoResults(), ProcessBumphunterResults(), and
  #   BuildPRcurve() functions.
}
\keyword{internal}
