% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_combine_dmrs_cpgs.R
\name{MergeDMRsWithCPGs}
\alias{MergeDMRsWithCPGs}
\title{Merge CPGs with DMRs}
\usage{
MergeDMRsWithCPGs(DMRs_df, CPGs_df, alpha = 0.05)
}
\arguments{
\item{DMRs_df}{A data frame with variables \code{dmr.start}, \code{dmr.end},
\code{dmr.chr} (e.g. \code{chr1}), and \code{dmr.pval}. This data frame
will be returned by the \code{extractRanges()} function from the
\code{DMRcate} package.}

\item{CPGs_df}{An annotation table that indicates locations of CpGs.
This data frame has CPG IDs as the rows with matching chromosome and
location info in the columns. Specifically, the columns are: \code{ILMNID}
 - the CPG ID; \code{chr} - the chromosome label; and \code{MAPINFO} -
the chromosome location. An example is given in the \code{cpgLocation_df}
data set.}

\item{alpha}{Significance level to select regions (with \code{dmr.pval} less
than the specified value). Defaults to 0.05.}
}
\value{
A data frame containing DMR info, CPG info, and the number of CPGs
}
\description{
An internal wrangling function for \code{RunDMRcate}
}
\examples{
  # Called internally by the StandardizeOutput() function.
}
\keyword{internal}
