% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_read_and_summarize_DMRcate_results.R
\name{ProcessDMRcateResults}
\alias{ProcessDMRcateResults}
\title{Process DMRcate Results Files}
\usage{
ProcessDMRcateResults(resultsDir, beta_mat, AclustCPG_df, verbose = TRUE)
}
\arguments{
\item{resultsDir}{The name of the directory where the DMRcate method results
are stored. This should match the directory name supplied to the
\code{resultsDir} argument of the \code{\link{WriteDMRcateResults}}
function.}

\item{beta_mat}{A beta value matrix for methylation samples from a
450k methylation array with CpG IDs as the row names and sample IDs as
the column names. An example is given in the \code{betaVals_mat} data set.}

\item{AclustCPG_df}{A data frame of beta values and CpG information for
clusters of CpGs over a 450k methylation array. The rows correspond to the
CPGs. The columns have information on the cluster number, chromosome,
cluster start and end locations, and the beta values for each subject
grouped by some clinical indicator (e.g. case v. control). An example is
given in the \code{startEndCPG_df} data set. This file also has information on true status of the clusters, via variable \code{status}, 
with values "positive" or "negative".}

\item{verbose}{Should the function print progress messages? Defaults to
\code{TRUE}.}
}
\value{
A data frame of model performance measures for the DMRcate method under
   each of the given parameter combinations applied to the data generated with different treatment effects
}
\description{
Given a directory of saved DMRcate results, as written by the
   \code{\link{WriteDMRcateResults}} function, import and summarize these
   data files.
}
\examples{
\dontrun{
  data("betaVals_mat")
  data("startEndCPG_df")

  dmrcateRes_df <- ProcessDMRcateResults(
    resultsDir = "DMRcate_results/",
    beta_mat = betaVals_mat,
    AclustCPG_df = startEndCPG_df
  )
}
}
