% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_RunDMRcate.R
\name{RunDMRcate}
\alias{RunDMRcate}
\title{Return Results from the \code{dmrcate} Function}
\usage{
RunDMRcate(betaVals_mat, labels_fct = factor(c(rep("Tumor", 7), rep("Normal",
  7))), cpgLocation_df, lambda_int, C_int, nCores = 1,
  dmr.sig.threshold = 0.05, min.cpgs = 5, genome = "hg19")
}
\arguments{
\item{betaVals_mat}{A matrix of beta values returned in the second entry of
the output from the \code{SimulateData} function.}

\item{labels_fct}{A factor vector of subject class labels. These should
match the observations contained in the columns of the \code{betaVals_mat}
matrix. Defaults to \code{factor(c(rep("Tumor", 7), rep("Normal", 7)))}}

\item{cpgLocation_df}{An annotation table that indicates locations of CpGs.
This data frame has CPG IDs as the rows with matching chromosome and
location info in the columns. Specifically, the columns are: \code{ILMNID}
 - the CPG ID; \code{chr} - the chromosome label; and \code{MAPINFO} -
the chromosome location. An example is given in the \code{cpgLocation_df}
data set.}

\item{lambda_int}{Gaussian kernel bandwidth for smoothed-function estimation
in the called \code{\link[DMRcate]{dmrcate}} function.}

\item{C_int}{Scaling factor for bandwidth in the internal call to the
\code{\link[DMRcate]{dmrcate}} function}

\item{nCores}{How many cores should be used to perform calculations? Defaults
to 1. Note that this function should be called from within the
\code{\link{WriteDMRcateResults}} function, which is already written in
parallel. Further note that the \code{DMRcate} package (as of version
1.16.0), does not support parallelization in Windows environments.}

\item{dmr.sig.threshold}{Significance level to select regions (with
\code{dmr.pval} less than the specified value) passed to the internal
\code{\link{StandardizeOutput}} function.}

\item{min.cpgs}{The minimum number of CPGs necessary to consider a result
significant, passed to the internal \code{\link{StandardizeOutput}}
function. Defaults to 5.}

\item{genome}{Reference genome for annotating DMRs with promoter overlaps,
passed to the \code{\link[DMRcate]{extractRanges}} function. Can be one
of \code{"hg19"}, \code{"hg38"}, or \code{"mm10"}. Defaults to
\code{"hg19"}.}
}
\value{
A list of two elements: a data frame of \code{dmrcate} results
   that have been standardized by the \code{\link{StandardizeOutput}}
   function and the computing time for the DMRcate method.
}
\description{
A wrapper function for the DMRcate method from the
   \code{DMRcate} package, called internally by the
   \code{\link{WriteDMRcateResults}} function.
}
\examples{
# Called internally by the WriteDMRcateResults() function.
\dontrun{
   data("betaVals_mat")
   data("cpgLocation_df")
   data("startEndCPG_df")

   treat_ls <- SimulateData(beta_mat = betaVals_mat,
                            AclustCPG_df = startEndCPG_df,
                            delta_num = 0.4,
                            seed_int = 100)
   class_fct <- factor(c(rep("Tumor", 7), rep("Normal", 7)))

   RunDMRcate(
     betaVals_mat = treat_ls$simBetaVals_df,
     labels_fct = class_fct,
     cpgLocation_df = cpgLocation_df,
     lambda_int = 500, C_int = 5
   )
}
}
