% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_simulatedata.R
\name{SimulateData}
\alias{SimulateData}
\title{Simulate Differentially Methylated Regions (DMRs) in Methylation Data}
\usage{
SimulateData(beta_mat, AclustCPG_df, delta_num, seed_int, betaCols_idx = 9:22,
  numEx_int = 7, numClusters_int = 500)
}
\arguments{
\item{beta_mat}{A beta value matrix for methylation samples from a
450k methylation array with CpG IDs as the row names and sample IDs as
the column names. An example is given in the \code{/data/betaVals_mat} data set.}

\item{AclustCPG_df}{A data frame of beta values and CpG information for
clusters of CpGs over a 450k methylation array. The rows correspond to the
CPGs. The columns have information on the cluster number, chromosome,
cluster start and end locations, and the samples grouped by some clinical indicator 
(e.g. case v. control). An example is given in the \code{/data/startEndCPG_df} data set.
This file can be generated by the file \code{/inst/1_Aclust_data_import.R}}

\item{delta_num}{The treatment size: a non-negative real number to add to
the beta values within randomly-selected clusters for the first  \code{numEx_int} samples. 
This artifically creates differentially-methylated regions
(DMRs).}

\item{seed_int}{The seed value passed to the \code{\link[base]{Random}}
function to enable reproducible results}

\item{betaCols_idx}{The column numbers of the \code{AclustCPG_df} data frame
in which beta values for each subject are stored.}

\item{numEx_int}{The number of samples in the first group. This
function assumes that samples in one group are contiguous columns of the
\code{AclustCPG_df} data frame.}

\item{numClusters_int}{The total number of randomly selected clusters, for which treatment effect \code{delta_num}
is then added}
}
\value{
A list with two elements:
   \itemize{
     \item{\code{simBetaVals_df}}{ A data frame of beta values after
       treatment effects were added, used for input for different DMR-finding
       methods. Note this dataset includes all the CpGs on the array.}
     \item{\code{simAclusters_df}}{ A data frame of the methylation values
       only for clusters identified by \code{Aclust} and indicator variable for whether treatment effects
       were added. Note this has only CpGs mapped to all the clusters found
       by the \code{Aclust} method.}
   }
}
\description{
This function generates a methylation dataset where treatment effects are added 
to beta values in one group of samples for some randomly selected co-methylated clusters.
}
\examples{
\dontrun{
   data("startEndCPG_df")
   data("betaVals_mat")

   SimulateData(beta_mat = betaVals_mat,
                AclustCPG_df = startEndCPG_df,
                delta_num = 0.4,
                seed_int = 12345)
}
}
