% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_summarize_DMR_results.R
\name{SummarizeResults}
\alias{SummarizeResults}
\title{Summarize Method Accuracy of Returned Results}
\usage{
SummarizeResults(cleanDMR_df, time_num)
}
\arguments{
\item{cleanDMR_df}{A data frame of joined predicted and true clusters, as
returned by the \code{\link{CleanResults}} function.}

\item{time_num}{The computing time for the given method and parameter
configuration, as stored in the second entry of the output list from the
\code{\link{WriteDMRcateResults}}, \code{\link{WriteBumphunterResults}},
or \code{\link{WriteProbeLassoResults}} functions.}
}
\value{
A data frame of model fit statistics for the given method to the
   data generated according to the given parameter configuration:
   \itemize{
     \item{\code{time}}{},
     \item{\code{FN}}{},
     \item{\code{FP}}{},
     \item{\code{TN}}{},
     \item{\code{TP}}{},
     \item{\code{power}}{},
     \item{\code{nPower}}{},
     \item{\code{FPprecis}}{},
     \item{\code{TPprecis}}{},
     \item{\code{precision}}{},
     \item{\code{nPrecis}}{},
     \item{\code{nCPG_q1}}{},
     \item{\code{nCPG_med}}{},
     \item{\code{nCPG_q3}}{}
   }
}
\description{
Given the output of the \code{\link{CleanResults}} (cleaned and
   joined predicted clusters to true clusters), return a data frame of the
   precision, power, elapsed time, and other summary statistics.
}
\examples{
  # Called internally by the ProcessDMRcateResults(), ProcessCombpResults(),
  #   ProcessProbeLassoResults(), and ProcessBumphunterResults() functions.
}
\keyword{internal}
