% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_standardize_and_summarize_Comb-p_results.R
\name{ProcessCombpResults}
\alias{ProcessCombpResults}
\title{Extract and Process Comb-p Results Files}
\usage{
ProcessCombpResults(resultsDir, beta_mat, AclustCPG_df, cpgLocation_df,
  dmr.sig.threshold = 0.05, min.cpgs = 5, verbose = TRUE)
}
\arguments{
\item{resultsDir}{The name of the directory where the Comb-p method results
are stored.}

\item{beta_mat}{A beta value matrix for selected methylation samples from a
450k methylation array with CPG IDs as the row names and sample IDs as
the column names. An example is given in the \code{betaVals_mat} data set.}

\item{AclustCPG_df}{A data frame of beta values and CpG information for
clusters of CpGs over a 450k methylation array. The rows correspond to the
CPGs. The columns have information on the cluster number, chromosome,
cluster start and end locations, and the beta values for each subject
grouped by some clinical indicator (e.g. case v. control). An example is
given in the \code{startEndCPG_df} data set.}

\item{cpgLocation_df}{An annotation table that indicates locations of CpGs.
This data frame has CPG IDs as the rows with matching chromosome and
location info in the columns. Specifically, the columns are: \code{ILMNID}
 - the CPG ID; \code{chr} - the chromosome label; and \code{MAPINFO} -
the chromosome location. An example is given in the \code{cpgLocation_df}
data set.}

\item{dmr.sig.threshold}{Significance level to select regions (with
\code{dmr.pval} less than the specified value) passed to the internal
\code{\link{MergeDMRsWithCPGs}} function.}

\item{min.cpgs}{The minimum number of CPGs necessary to consider a result
significant. Defaults to 5.}

\item{verbose}{Should the function print progress messages? Defaults to
\code{TRUE}.}
}
\value{
A data frame of model fit statistics for the Comb-p method under
   each of the given parameter combinations to the data generated for each
   design configuration
}
\description{
Given a directory of saved Comb-p results, as \code{.RDS} files,
   import, standardize, and summarize these data files.
}
\examples{
\dontrun{
  data("betaVals_mat")
  data("startEndCPG_df")
  data("cpgLocation_df")

  combpRes_df <- ProcessCombpResults(
    resultsDir = "DMRcate_results/",
    beta_mat = betaVals_mat,
    AclustCPG_df = startEndCPG_df,
    cpgLocation_df = cpgLocation_df
  )
}
}
